+++ app/models/product.js
@@ -19,7 +19,7 @@ newitemcost : DS.attr('number'),
 
 
 computedtotal_retail_intialstocklevel: Ember.computed('retailprice' , 'initialstocklevel', function() {
-    return parseInt(this.get('initialstocklevel')) * parseInt(this.get('retailprice'));
+    return parseInt(this.get('initialstocklevel')) * parseFloat(this.get('retailprice'));
   }),
