+++ app/models/stockadjustmentitem.js
@@ -13,7 +13,7 @@ export default DS.Model.extend({
   }),
 
   computedtotal: Ember.computed('product' , 'quantity', function() {
-    var total =   parseInt(this.get('product.retailprice')) * parseInt(this.get('quantity'));
+    var total =   parseFloat(this.get('product.retailprice')) * parseInt(this.get('quantity'));
      return total;
   }),
 });
